repo-updater: Trim trailing / from github repo URL
Previously we assumed that the github API would never return a repo URL which
had a trailing `/`. This assumption made the code harder to verify correct,
even if in practice it works. So we introduce a helper function to join the
URL.

Note: We can't use path.Join since it will path normalize the URL scheme. We
also can't use a parsed URL since `{` and `}` in the template will be escaped
when formatting the URL.