repo-updater: Parameterise http.Client of GithubSources
This change-set parameterises the http.Client used by GithubSources to
enable use to more easily test these Sources without huge setup and
run-time costs.

It introduces a helper package called httpcli which faciliates the
construction of a fully featured http.Client, wrapped with Middleware
when necessary.

This is a pre-cursor PR to #2649 which will need extensive tests on
GithubSource.