Clean up repository inventory code (#2669)
* remove internal API for repo inventory

These are no longer used. They were previously used by the indexer, which triggered indexing of defs and refs data for cross-repository code intelligence. This indexer was removed prior to Sourcegraph 3.0.

* move Go package ./pkg/inventory to underneath ./cmd/frontend/...

Only the frontend uses the inventory (which is the listing of languages used in a repository), so it can and should become an internal package under ./cmd/frontend.