repo-updater: Debug log and trace repoLookup (#2679)
We are having flaky behaviour. This will help us narrow down why we are
experiencing it.

---

Conflicts:
	cmd/repo-updater/repoupdater/server.go

I (@stephen) manually addressed this conflict by backporting the change.
The only part of this change I didn't feel confident in making was adding
back this log line 201:

```Go
tr.LazyPrintf("authorative: %s", get.kind)
```

This was because get.kind requires refactoring the getfn's into a structured
type like in `master`, but this is a larger change and the change is otherwise
easily verifiable (i.e. I feel comfortable committing this change without that
refactor, but would otherwise need @keegancsmith's review). If this log info is
important, though, we can of course backport it.