Massively reduced the number of code host API requests Sourcegraph performs (#2630)
* cmd/frontend/backend: VERY substantially reduce code host rate limit consumption

TL;DR

Prior to this change, for every search result you viewed we would perform 4 code
host API requests and this led to us running out of our rate limit rapidly. It
causes:

- [Search results to take a very long time to load when using Bitbucket Server.](https://github.com/sourcegraph/sourcegraph/issues/2618)
- Links to the GitHub repository, etc. to not appear on repo pages.
- And likely many, many, more issues.

After this change, we massively reduce the number of code host API requests we
perform because it turned out in 99% of cases (not just when searching) we were
doing them for literally no reason due to unfiled technical debt.

Detailed explanation

Repository revision resolutions (`ResolveRev` and `GetCommit`) occur very often
in our codebase. For example, every search result you view makes two of these
calls.

Prior to this change, each revision resolution would consume two code host API
requests (which are heavily rate limited) due to the fact that they must
perform a repo-updater `/repo-lookup` request in order to determine what the
remote URL for the repository is when updating the gitserver repo mirror.

We already knew this was costly, and in fact that is why there is a
`CachedGitRepo` variant of `GitRepo` in this same file. Basically, if you don't
need the remote URL we can spare ourselves tons of work and, most importantly,
code host API reservations.

But revision resolution wasn't able to use `CachedGitRepo` because we operated
under the assumption that when resolving a revision you _do_ want the repository
to update. For example, when resolving `master` or `mybranch` you want the
repository updated to reflect the latest state of those branches.

Times changed, repo-updater changed how we did things, and we started operating
with more explicit requests to update repositories via `repoupdater.DefaultClient.EnqueueRepoUpdate`.

At some point, gitserver became not responsible for updating repositories _unless_
the rev in question was not known to it (i.e. `mybranch` may be stale if already
known to gitserver until repo-updater updates it, but gitserver will explicitly
update the repository if `mybranch` is unknown to gitserver). You can see this logic clearly in gitserver's [`ensureRevision` function](https://sourcegraph.com/github.com/sourcegraph/sourcegraph@e1f88f6dfc5cbdc6cac0754cf4eeff1492609253/-/blob/cmd/gitserver/server/server.go#L1275-1276).

What this means is that despite the fact that we always use a `GitRepo` for
`ResolveRev` and `GetCommit` requests, in 99% of cases gitserver already has
the rev and will not update it, _but we'll still ask repo-updater to do all of
that work and make those code-host requests just cuz'.

After this change, the user observed behavior remains the exact same: gitserver
doesn't update revs it is already aware of (but note our UI does ask it to in
specific locations via `EnqueueRepoUpdate`). We just don't ask repo-updater to
do all that work for nothing anymore.

Helps #2618

* Revert "cmd/frontend/backend: VERY substantially reduce code host rate limit consumption"

This reverts commit 9be47c082988308e18be9cc41c85f87884c04d8c.

* pkg/vcs/git: allow GetCommit to take a remoteURLFunc, similar to ResolveRevision

* cmd/frontend/backend: make ResolveRev/GetCommit use of remoteURLFunc

* CHANGELOG: mention this PR

* pkg/vcs: refactor retrying logic into shared commandRetryer type

* fix linter error

* rm unused NoEnsureRevision