show Go package "used by" counts including repository subpackages (#2604)
* show Go package "used by" counts including repository subpackages

Previously, our "used by" badge (for Go packages) would only show the number of other packages using a repository's *top-level* package, not its subpackages. Now the "used by" badge counts users of its subpackages.

The count sums together the user counts of all of the repository's subpackages. This overcounts the number of users, because if another project uses multiple subpackages in this repository, it is counted multiple times. This limitation is now documented and will be addressed in the future.

Fixes some of https://github.com/sourcegraph/sourcegraph/issues/849 (does not implement actually showing the packages that use it).

* document double-counting

* check for repo enabled

* use path.Join

* fix docstring

* cache using redis, not in memory