add settings action for search.contextLines (#2755)
Adds a button to change the `search.contextLines` value on the page where you edit the global/org/user settings JSON. Improves the situation in #2754.

There is a bug in our quick action buttons where the Monaco editor's selection after running the quick action is incorrect in some cases (e.g., if the edit it applied also added a comma to the preceding line). This is annoying but is not critical.