Fix various minor code discussions bugs, make tweaks (#2791)
* rm space

* clean up code discussions panel

- Make only the title and comment count clickable, not the full row. This is (to some people) a slight UX regression, but it fixes many issues. This is how GitHub Issues' UI works.
  - fix #1072
  - fix #1215
  - fix #1368
- Tweak wording of rows.
- Remove unnecessary tooltips.
- Remove CSS that can be expressed simply in Bootstrap.
- Fix padding for "N code discussions." summary text.

* remove discussions explore section

It contained duplicative impl code and diverged from how discussions were shown in the panel. Because it was not widely used in the explore page, it's not worth keeping it there.

fix #1100

* remove unnecessary code discussions tooltips

* select correct line (fix off-by-1 error) when viewing thread

fix #428 (for the selected line)

* fix issue where navigating in code discussions panel was no-op

* fix linking directly to comments and jitter when clicking on thread

Because the inlineURL now includes the L123& part of the URL, there is no longer jitter when clicking on a thread in the discussions panel (where the "Definitions" and "References" panel labels briefly go away).

* navigate to comment when copying link

This makes the icon work more like the user (IMO) expects.

* remove code discussions URL hash keys when closing panel

* strip needless trailing = (as in #L23=) when closing panel

This trailing = is there because `L23` is interpreted as a key with an empty string value. It is valid in HTML form encoding to omit `=` and the value in this case, but for whatever reason, `hash.toString()` keeps `=`.

* remove need to pass through authenticatedUser in code discussions React components

* add TODO about fixing bug when deleting 1st comment in thread

https://github.com/sourcegraph/sourcegraph/issues/429