Add new fileLineContainer code view for github
Fixes #2857

Github.com's DOM changed, and the `'.file'` selector we used to rely on is obsolete.

This adds a new `CodeView` to handle the new Github DOM (but preserves the old selector because it is still useful on some PRs, and probably on GHE).

This new code view does not reuse the existing `createCodeViewToolbarMount()`, because it also relies on obsolete selectors (`'.file-actions'`).