docs: updates for the GitHub external service
Inspired by #2895 and learning how repo querying works.

I've aimed to more fully explain the repository syncing design/strategy, as well as adding important points such as which token to use. 

This is not to replace the schema docs, but provide a high level understanding so the schema is more easily understandable.