repo-updater: Prevent HTTPS redirect in github source tests
It seems when we redirected we wouldn't set the Authorization header, leading
to our requests actually being unauthenticated. I noticed this problem when
adding a test case which returned a large result set (1000+). The test was
failing to record due to rate limits, since the requests were actually
unauthenticated.

I chose to just make a specific wrapper for github-proxy since this is the
only use of it, it will likely remain the only use of it, and making it
generic is less clear / unneccessary given the previous points.