repo-updater: Safer EnabledStateDeprecationMigration
This commit makes the EnabledStateDeprecationMigration delete and enable
the disabled repos it excludes in external services.

Together with the UI changes that prevent admins disabling repos, as well as the
new enabled default, this ensures that they won't be touched a second
time a migration is ran.

The alternative of keeping track of which migrations already ran proved
to be more work than expected, so I'm putting this option forward for
review.