pkg/tracer: allow Jaeger to be configured via standard JAEGER_* env vars
Prior to this change it was not possible to configure the Jaeger client using
[the standard JAEGER_* env vars](https://github.com/jaegertracing/jaeger-client-go/blob/v2.14.0/README.md#environment-variables).

After this change, behavior remains the same for all Jaeger users except
JAEGER_* env vars are now respected.

Without this change, it is impossible to configure Sourcegraph services to
communicate with a `jaeger-agent` running on the same host machine in the case
of deploy-sourcegraph-docker because the loopback interface does not work to
communicate with same-host containers there (while it does on Kubernetes), i.e.
we must use their actual domain name like `jaeger-agent`.

Helps #2873