consistently use view component type "CodeEditor" not "textEditor" (#3267)
In sourcegraph.d.ts, the `CodeEditor#type` value is `CodeEditor`, but in context expressions and in the impl, the `type` is `textEditor`. This commit updates everywhere to use `CodeEditor` for consistency.

All extensions with context expressions using the expression `component.type` need to be updated to check for `"CodeEditor"` not `"textEditor"` (and should be backcompat with older Sourcegraph versions, so both values should be supported and treated identically).