fix hasProvidersForActiveTextDocument test (#3276)
The `hasProvidersForActiveTextDocument` method returns false when there are no selections. The test name indicates that it should be testing when there IS a match, but the test code itself checked for false and set up a scenario where there was no match. This fixes the test code to match its name.