all: Propogate opentracing headers
In an old refactor were we switched to using ctxhttp more, we regressed on
propogating the opentracing headers across RPC calls. This was due to
nethttp.TraceRequest modifying the request's context. However, we would then
do ctxhttp which would update the context before the call to
nethttp.TraceRequest. This lead to nethttp.Transport not finding the request
span to inject.

Also changed in this commit:

- Upgrade opentracing libraries
- Remove unneeded span from gitserver.exec