api: Ensure WaitForFrontend blocks
This renames `RetryPingUntilAvailable` to `WaitForFrontend` which
replaces the previous package level functions of the same name.

That function was previously assumed to be blocking until the frontend
was available, but it instead timed out after 5 seconds.

We were seeing problems in development where repo-updater failed because
migrations hadn't been ran yet. This meant that we weren't waiting for
the frontend until it was actually ready.

This commit changes that.

Part of #2025