only show Sourcegraph icon, no text, in "View file" buttons (#3279)
The browser extension's buttons on file and diff headers on code hosts currently read "View file" and "View file diff". These take up a lot of width (which creates jitter and sometimes means that the buttons can require more than 1 line, which causes even more UI jitter).

It is almost certainly clear enough to users that the buttons mean "View this thing on Sourcegraph". The buttons have tooltips that describe what they do, so users are unlikely to be confused.

## After

![image](https://user-images.githubusercontent.com/1976/55696825-8c367100-5973-11e9-83c8-91908c2851d5.png)

![image](https://user-images.githubusercontent.com/1976/55696835-95bfd900-5973-11e9-9057-e0038bdb1f8a.png)

---

## Before
![image](https://user-images.githubusercontent.com/1976/55696858-b38d3e00-5973-11e9-8ca0-5f193cabe55b.png)


![image](https://user-images.githubusercontent.com/1976/55696851-a96b3f80-5973-11e9-84e9-ec9b5dd1b4eb.png)