repo-updater: Idempotent scheduler operations
This commit makes the additive operations of the scheduler
internal queues idempotent.

This simplifies usage of the scheduler and fixes a bug in
`UpdateFromDiff`. This bug manifested when already deleted repos
remained deleted (in the `Unmodified` diff field) but they were still
being added instead of removed from the schedule.

This could happen everytime repo-updater would startup with already
deleted repos existing in the database.