new extension API (LinkPreviewProvider): show link previews on GitHub issues
Adds a new extension API `sourcegraph.content.registerLinkPreviewProvider` that can be used to annotate links in a GitHub issue (or other comment) with hover and after content. This can be used, e.g., to show preview cards for JIRA issues (not just other GitHub issues), customer names for private CRM links (to track which customers care about an issue), thumbnails of designs on Figma/Zeplin, etc.

This is only implemented for GitHub now, but the hooks in the `CodeHost` interface can be used to add support for other code hosts.

The extension API is marked as experimental for now.