web: Add external service cards to repo settings page
We include links to the external service settings for a repository. This is
only shown on the repository settings page, which is only visible to
admins. The list of external services returned from the backend is non-empty
iff the repository is managed by our new syncer. For repositories not managed
by our new syncer, we still show the disable and delete repository actions.