github: Fail if search has incomplete results
Our new syncer requires a source to return all repositories it is configured
to return. A repository not being listed by a source will be deleted by the
syncer. So in the case of incomplete results a repository will be
soft-deleted (and then likely be undeleted the next sync). Instead we just
fail the search if GitHub returns incomplete results.