debugserver: Add zoekt services to debug server (#3360)
* dev: Use the same port for zoekt-webserver as server image

We like to keep the ports consistent between the two.

* Add zoekt debug ports to site admin instrumentation proxy

We have a convenience page at http://sourcegraph.example.com/-/debug/ for site
admins to view individual service's debug pages. This adds zoekt to the list.

* debugserver: Support setting a path to link to

For zoekt-webserver we want to link to the debug page since it presents its
search interface at `/` without any links to the debug page.