repo-updater: Syncing improvements for OTHER external services
This commit integrates the OTHER external service kind with the new
Syncer. This was required to do in order to make the scheduler handle
renames because the `OtherExternalServicesSyncer` emitted
`protocol.RepoInfo` which doesn't have an ID.

Populating it's ID would require changes that we wouldn't be using in
the long run, so I gave it a shot to integrate this with the new Syncer.

Notably, the syncer won't handle renames nor deletes of repos of this
kind since we don't have a code host API as a source of truth.

Part of #2025