add extension API model service, separate models and editors (#3322)
* extract EditorService#editors into subscribable, next, and value for easier future refactoring

This will let us make the subscribable merge in data from other sources (while still keeping the next/value APIs only dealing with normalized data).

* add extension API model service, separate models and editors

Previously, document contents and metadata were stored in the editor service. This was simple, but it made it hard to manage documents separately from editors, which is helpful when editors and documents are no longer 1-to-1 (as in the case of diff views, a future desired feature).

* remove need for most explicit syncing and double-syncing, simplify ModelService API

addresses https://github.com/sourcegraph/sourcegraph/pull/3322#discussion_r273438174

* fix code intel test

* rm comment

* fix test

* fix jsdoc

* add ModelService#updateModel

This allows callers to update the text content of a model. It is not currently used, but it will be used in the future (initially for supporting completions on GitHub issue textareas).