combine user subscriptions tab with settings tab
This makes the subscriptions area just another item in the user settings sidebar. Also updates URLs to subscriptions (but old URLs are supported with redirects).

The new preferred URL for unauthed users to price out a plan is `/subscriptions/new` not `/users/subscriptions/new`. (The old one still works via a redirect.)