combine settings & account pages for users and orgs (#3397)
* add back org account area

It is currently empty, but this is an intermediate commit along the way to combine the org profile and settings in a single tab in the org area.

This partially reverts commit f87063aa7fa5bc03458fef63808734fd6abef9ef.

* mv web/src/org/{account => settings}

* mv web/src/org/settings/*{Account => Settings}*

* rename Org{Account => Settings}*

* move fetchViewerSettings

* mv web/src/user/{account => settings}

* mv web/src/user/settings/*{Account => Settings}*

* rename User{Account => Settings}*

* combine user settings and account area into 1 tab in user area

* combine org settings and account area into 1 tab in org area

* fix settings top margin

* explain subscriptions

* combine user subscriptions tab with settings tab

This makes the subscriptions area just another item in the user settings sidebar. Also updates URLs to subscriptions (but old URLs are supported with redirects).

The new preferred URL for unauthed users to price out a plan is `/subscriptions/new` not `/users/subscriptions/new`. (The old one still works via a redirect.)

* only set up user external accounts settings route when auth experiment is on

* update name