repo-updater: Enable new repository syncer by default
The only environment we don't want to be running the feature flag yet is
Sourcegraph.com. I have just confirmed that it has
`SRC_SYNCER_ENABLED: false` on the repo-updater pod in production.

We are deferring the work to special case Sourcegraph.com until early in the
next milestone so that we can focus on more pressing issues before branch cut
tomorrow.