repo-updater: Adjust rate limit sleep heuristic for search
GitHub source which relied on repositoryQuery would be very slow, normally
leading to a timeout. This was due to the very low rate limits and the "sleep"
heuristic being based on the normal GitHub rate limit. Essentially the sleep
operation would always be 10s. This commit instead makes it possible for us to
adjust the heuristic for sleeping for 0s from 500 left to 8 left.