repo-updater: Adjust sleep rate limit heuristic for GitHub Search API (#3400)
GitHub source which relied on repositoryQuery would be very slow, normally
leading to a timeout. This was due to the very low rate limits and the "sleep"
heuristic being based on the normal GitHub rate limit. Essentially the sleep
operation would always be 10s. Instead we only rely on the heuristic if there
is less than 5 requests left.