new extension API for completion items (#3392)
* (temporarily) remove username mention completions in comment thread inputs

These will be re-added using an implementation based on extensions. This is an intermediate commit to make the diffs easier to review.

* new extension API for completion items

The new `sourcegraph.languages.registerCompletionItemProvider` API lets extensions register providers to suggest completions for a cursor position in an editor.

This is supported in code discussions comment input fields, GitHub issue edit/create fields, and on an experimental `/snippets` page on the web app.