make /subscriptions/new not 404 (#3420)
The new URL https://sourcegraph.com/subscriptions/new for unauthed users to price out a plan was merged in #3397, but I forgot to add a Go route entry for it. This meant that it would 404 and the tab title would briefly read "Not found". These are minor issues that this commit fixes.

(The URL used to be https://sourcegraph.com/user/subscriptions/new, which was already handled by the existing Go user route.)