fix completion widget position in scrolled textarea
Despite https://github.com/component/textarea-caret-position/blob/master/README.md claiming it supports scrolled textareas ("no problem getting the correct position when the input text is scrolled (i.e. the first visible character is no longer the first in the text)"), it was not working. If the textarea was scrolled down, the completion widget would be displayed well below the cursor (by the amount that the textarea was scrolled down).

fix #3424