WIP: MessageEventLike
Attempts to address https://github.com/sourcegraph/sourcegraph/pull/3434#discussion_r275951925

Unfortunately, our code is too intertwined with comlink's API (not in a bad way, just in a way that makes it harder to use a synthesizable type). So I am going to see if we can make comlink's API use MessageEventLike instead of introducing it in our code.