repo-updater: BitbucketServerUsernameMigration
This commit implements a best effort migration that extracts the
username from the `url` field of Bitbucket Server external service
configurations and sets in the `username` field if missing or empty.

The `username` field became required in #3299 because it was always
actually required for things to fuction properly.

If there are configs without the `username` field set not the `user`
part of the `url` set, the admins will need to intervene to fix their
configs manually.

Fixes #3592