repo-updater: Do not migrate deleted repos to `(repos|projects)` list (#3598)
This commit is a proposed solution to #3588. It changes the
`EnabledStateDeprecationMigration` so that it doesn't migrate deleted
repos, nor repos that would have been deleted on the first syncer run,
to an explicit include list (`repos` in Github and Bitbucket Server,
`projects` in Gitlab).

There were reports of confusion around these `repos` lists being
populated in all external service configurations. The explanation for
this is that we previously didn't have a link between repos and the
external services that yielded them. Because this migration could not
know which external service a repo came from, it added these include and
exclude lists to ALL external services of the same kind.

While this safe to do for `exclude`, we know learned that it wasn't for
`repos|projects` due to these requests to code hosts inquiring about deleted
repos consuming rate limit quota. Having this `repos` list present in every
active external service that shared an account / token, accelerated
the exhaustion of the rate limit quota.

Fixes #3588