repo-updater: add dumb mitigation for duplicate repo issue #3680
Today, #3680 prevents a customer from upgrading successfully to 3.3 due to the
fact that (I believe) when this duplicate repository state occurs it blocks all
other repositories from updating.

While we gather more data and debug that issue further, I'd like to unblock
that customer, so I've implemented some dumb mitigation here that allows the
system to work as expected.

Additionally, this change includes extra tracing which will help us debug
whether or not the DB index is potentially corrupted (i.e. we violated the
unique repo constraint somehow) OR if it is indeed a logic bug.

Once we've resolved the underlying issue we should of course revert this
change.

After merge, I will ship this in a patch release.

Helps #3680