repo-updater: Upsert relies on batch ID
Currently we duplicate logic in NewDiff and Store were we attempt to
deduplicate repos which conflict on Name or ExternalService. Instead we can
simplify the upsert logic and rely on NewDiff not returning duplicates.

This does mean that a partial Diff (not every kind) could conflict on
name. However, name's should not conflict across kinds excluding users who use
our broken repositoryPathPattern. Additionally letting the database take care
of this conflict would lead to surprising changes to the table.

I believe the original code needed to handle the cases due to us not running
diff on every repository.

Note: This simplifies the FakeStore. Additionally we remove the unused
GetRepoByName fake.