repo-updater: Exclude patterns for Bitbucket Server
For some customers it is impractical to enumerate the valid repository patterns,
since the list is large and constantly changing. However, they just want to
exclude a specific project. This commit adds the more general ability to exclude
based on a regular expression for Bitbucket Server. Its approach can be extended
to other code hosts as needed.