httpcli: Separate Doer and Client methods in Factory
This commit renames the `NewClient` method in `httpcli.Factory` to
`Doer` and introduces a new `Client` method that returns a concrete
`*http.Client` with all options applied to it but without middleware.

This is to integrate with packages that can't easily be made to work
with a `Doer` interface like the AWS client package.