repo-updater: Do not delete repos in EnabledStateDeprecationMigration
We would only run the delete step in the migration. This meant that in the case
of us changing the external_id we would delete repositories and only re-add them
once the syncer runs. This lead to a customer having all repositories deleted =>
our gitserver janitor job deleting all clones. This caused downtime due to the
slow time it took to reclone everything.

Alternatively we could do a full upsert from the diff.