repo-updater: Simplify EnabledStateDeprecationMigration
This commit simplifies the `EnabledStateDeprecationMigration`.

It removes old complexity associated with the longer valid requirement adding
deleted repos to an include list as well as removing the need to source
repos at all. Since all repos in a Diff are Enabled, we know that we can
simply access the disabled stored repos, add them to the relevant
exclude lists and delete them.