repo-updater: Add AWSCodeCommitSetBogusGitCredentialsMigration (#3918)
* repo-updater: Add AWSCodeCommitSetBogusGitCredentialsMigration

PR #3878 introduced a new, *required* setting for AWS CodeCommit
external services: `gitCredentials`. Without setting this, the config is
invalid and the repo-updater wouldn't boot up, because it failed in the
`EnabledStateDeprecationMigration` and retried it indefinitely.

That makes the upgrade for users needlessly hard:

- a user cannot add the `gitCredentials` setting to the config *before*
  upgrading, because that would make the config invalid
- *after* upgrading the user *can* add `gitCredentials` to the AWS
  CodeCommit but syncing the new config to repo-updater would fail since
  it's busy retrying the `EnabledStateDeprecationMigration`
- the user would have to manually restart repo-updater to pick up the
  updated and valid config

This commit remedies that problem but introducing a migration that runs
automatically when repo-updater boots up. This migration makes the
config "valid" by adding bogus `gitCredentials` that allow repo-updater
the and the rest of the system to boot up but also signal to the user
that they need to be changed.

* repo-updater: Remove needless whitespace

Co-Authored-By: Keegan Carruthers-Smith <keegan.csmith@gmail.com>

* Remove debug statement