repo-updater: Optimise EnabledStateDeprecationMigration
This commit optimises the `EnabledStateDeprecationMigration` so that it
only sources repos from the code host the first time it runs, when there
are still disabled repos stored in the database.

In subsequent runs, sourcing doesn't occur, because the disabled repos
have been already deleted and added to the exclude lists of the
relevant external services.

This makes repo-updater startup much faster.

Supersedes #3912