frontend: Remove Enabled field from types.Repo
We remove this field from all internal reads. We adjust all reads of the field
to act as if it was true. Additionally we update the database listing code to
always have the conditional `enabled = true`. Once the first sync of the new
syncer has run, all rows in the table will have enabled = true.