repo-updater: Increase batch when fetching Bitbucket Server repos
A customer reported performance issues when saving their external
service configuration for Bitbucket Server (which contains 10s of
thousands of repos)

I successfully replicated the issue locally by generating 20k artificial
Bitbucket Server repos in my local Bitbucket server instance and could
narrow the problem down to the syncer, which took several minutes to
fully fetch all 20k repos.

We're already using batching when fetching BB Server repos, but the
batch size is 100. (see [0])

According to the docs (see [1]) the maximum batch size is set by the
site administrator of the BB Server instance and the the default value
for `page.max.repositories` is `1000` (see [2])

So in this commit we set the default batch size to 1000.

[0]: https://sourcegraph.com/github.com/sourcegraph/sourcegraph/-/blob/cmd/repo-updater/repos/bitbucketserver.go#L269:29
[1]: https://docs.atlassian.com/bitbucket-server/rest/5.7.0/bitbucket-rest.html#paging-params
[2]: https://confluence.atlassian.com/bitbucketserver/bitbucket-server-config-properties-776640155.html#BitbucketServerconfigproperties-Paging