frontend: Only do github.com prefix search optimization on dot-com (#4103)
We have a heuristic which translates repo queries with the prefix `github.com`
into `^github.com`. This was likely added with the sourcegraph.com use case in
mind (3.5mil repos in the table and they all start with github.com). However,
for customers github.com can appear in other parts of the repo string. As such
we only use this optimization on Sourcegraph.com

This commit additionally includes some extra checks to ensure our include
pattern parsing doesn't make the same mistake.