support PersonLink elements without a separate displayName
This makes it possible to render a `<PersonLink />` when all you have is a `GQL.IUser`. There is no good reason for requiring a `displayName`. The only reason it was required was to avoid a case where the component's API implied that you could render it with neither a displayName nor a user. The new component API prevents that.