remove focus ring around clicked icons
When an icon is focused, such as when it's in a button, it acquires the standard OS focus ring (in my case, a blue border). This is undesirable.

For example, if you click-and-hold on the lock icon on a code discussion comment header and then move your mouse away, the focus ring displays on the icon, even though it is not a button and should not have a focus ring.