cleaner base CSS for <hr/>
This lets us remove overrides for its styles and makes it more visually appealing. Previously it was black (on the light theme) or white (on the dark theme), which was too much contrast. Now it is the border color, which is much subtler.

In cases where we actually want a lot of vertical margin, this commit adds that explicitly.