search: Handle ()() in autofix
Noticed this flakey test failure in CI. It is a legitimate issue were we do
not properly escape ()(). It escaped it as \(\)(). This is due to the
replacement regex checking for the preceding character being \, which means
when ReplaceAll runs after replacing the first () the regex won't match since
`^` means beginning of string, not match.

I wasn't sure how to make this happen in a regex, so wrote a function
implementing the functionality.